Fix body type from u64 to u32

+++ compiler/src/generate.rs
@@ -26,7 +26,7 @@ impl Display for Function {
 
 #[derive(Debug, Clone)]
 enum Instruction {
-    Ret(u64),
+    Ret(u32),
 }
 
 impl Display for Instruction {
